+++ tests/result/qprogram/test_qprogram_results.py
@@ -1,0 +1,102 @@
+""" Test Results """
+import json
+
+import numpy as np
+import pytest
+
+from qililab.result.qprogram.qblox_measurement_result import QbloxMeasurementResult
+from qililab.result.qprogram.qprogram_results import QProgramResults
+from qililab.result.qprogram.quantum_machines_measurement_result import QuantumMachinesMeasurementResult
+
+
+@pytest.fixture(name="qblox_qprogram_results")
+def fixture_qblox_qprogram_results():
+    """fixture_qblox_result_noscope
+
+    Args:
+        dummy_qrm (Pulsar): _description_
+
+    Returns:
+        _type_: _description_
+    """
+    results = QProgramResults()
+    results.append_result(bus="readout", result=QbloxMeasurementResult(raw_measurement_data={"abc": 123}))
+    return results
+
+
+@pytest.fixture(name="quantum_machines_qprogram_results")
+def fixture_quantum_machines_qprogram_results():
+    """fixture_qblox_result_noscope
+
+    Args:
+        dummy_qrm (Pulsar): _description_
+
+    Returns:
+        _type_: _description_
+    """
+    results = QProgramResults()
+    results.append_result(
+        bus="readout", result=QuantumMachinesMeasurementResult(I=np.linspace(0, 10, 11), Q=np.linspace(90, 100, 11))
+    )
+    return results
+
+
+class TestsQProgramResult:
+    """Test `QbloxQProgramResults` functionalities."""
+
+    def test_append_result_method_with_quantum_machines(self, quantum_machines_qprogram_results: QProgramResults):
+        """Tests the instantiation of a QbloxQProgramResult object.
+
+        Args:
+            qblox_result_scope (QbloxQProgramResult): QbloxQProgramResult instance.
+        """
+        assert "readout" in quantum_machines_qprogram_results.results
+        assert len(quantum_machines_qprogram_results.results["readout"]) == 1
+
+        quantum_machines_qprogram_results.append_result(
+            "readout", result=QuantumMachinesMeasurementResult(I=np.linspace(0, 10, 11), Q=np.linspace(90, 100, 11))
+        )
+        assert len(quantum_machines_qprogram_results.results["readout"]) == 2
+
+        quantum_machines_qprogram_results.append_result(
+            "another_readout",
+            result=QuantumMachinesMeasurementResult(I=np.linspace(0, 10, 11), Q=np.linspace(90, 100, 11)),
+        )
+        assert "another_readout" in quantum_machines_qprogram_results.results
+        assert len(quantum_machines_qprogram_results.results["another_readout"]) == 1
+
+    def test_append_result_method_with_qblox(self, qblox_qprogram_results: QProgramResults):
+        """Tests the instantiation of a QbloxQProgramResult object.
+
+        Args:
+            qblox_result_scope (QbloxQProgramResult): QbloxQProgramResult instance.
+        """
+        assert "readout" in qblox_qprogram_results.results
+        assert len(qblox_qprogram_results.results["readout"]) == 1
+
+        qblox_qprogram_results.append_result(
+            "readout", result=QbloxMeasurementResult(raw_measurement_data={"def": 456})
+        )
+        assert len(qblox_qprogram_results.results["readout"]) == 2
+
+        qblox_qprogram_results.append_result(
+            "another_readout", result=QbloxMeasurementResult(raw_measurement_data={"def": 456})
+        )
+        assert "another_readout" in qblox_qprogram_results.results
+        assert len(qblox_qprogram_results.results["another_readout"]) == 1
+
+    def test_serialization_method(self, qblox_qprogram_results: QProgramResults):
+        """Test serialization and deserialization works."""
+        serialized_dictionary = qblox_qprogram_results.to_dict()
+        assert "type" in serialized_dictionary
+        assert "attributes" in serialized_dictionary
+
+        deserialized_qp = QProgramResults.from_dict(serialized_dictionary["attributes"])
+        assert isinstance(deserialized_qp, QProgramResults)
+
+        again_serialized_dictionary = deserialized_qp.to_dict()
+        assert serialized_dictionary == again_serialized_dictionary
+
+        as_json = json.dumps(again_serialized_dictionary)
+        dictionary_from_json = json.loads(as_json)
+        assert serialized_dictionary == dictionary_from_json